repliq-backend/review-processing-py/requirements.txt
fastapi
uvicorn
google-adk
litellm
pymongo
prefect
opentelemetry-api
opentelemetry-sdk
loguru
pytest==7.4.4
pytest-cov
pytest-asyncio